Refactor bevy_renderer into modular submodules
Split the large bevy_renderer.rs file into multiple focused submodules: camera.rs, components.rs, metrics.rs, mod.rs, renderer.rs, resources.rs, scene.rs, sync.rs, textures.rs, and types.rs. This improves maintainability and organization by separating camera logic, components, metrics, resources, and other functionality into dedicated files.
